03_litellm_openai_agent/pyproject.toml
[project]
name = "03-litellm-openai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents[litellm]>=0.2.10",
    "python-dotenv>=1.1.1",
    "rich>=14.1.0",
]
